Remove `feature(specialization)`
