+++ src/arithimpl/mod.rs
@@ -1,5 +1,6 @@
 //! Supported underlying libraries for arbitrary precision arithmetic.
 
+mod defaultimpl;
 pub mod gmpimpl;
 pub mod rampimpl;
 pub mod traits;
